change int16 to int32 to meet tpu require

+++ readdata24.py
@@ -186,7 +186,7 @@ class DataSpeech():
 		while True:
 			X = np.zeros((batch_size, audio_length, 200, 1), dtype = np.float)
 			#y = np.zeros((batch_size, 64, self.SymbolNum), dtype=np.int16)
-			y = np.zeros((batch_size, 64), dtype=np.int16)
+			y = np.zeros((batch_size, 64), dtype=np.int32)
 			
 			#generator = ImageCaptcha(width=width, height=height)
 			input_length = []
